requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
discord.py==2.6.3
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
